feat(user): add update user and password use cases with tests
Add new use cases for updating user information and password, including comprehensive test coverage.
Move use cases to domain/user directory and update imports accordingly.
