+++ src/domain/user/use-cases/__test__/update-user-password.test.ts
@@ -1,7 +1,7 @@
-import { UpdateUserPassword } from "../update-user-password";
-import { UpdateUserDto } from "../../user/dtos/update-user.dto";
-import { UserEntity } from "../../entities/user.entity";
-import { AuthUserEntity } from "../../entities/auth-user.entity";
+import { UpdateUserPassword } from "@/domain/user/use-cases/update-user-password";
+import { UpdateUserDto } from "@/domain/user/dtos/update-user.dto";
+import { UserEntity } from "@/domain/entities/user.entity";
+import { AuthUserEntity } from "@/domain/entities/auth-user.entity";
 import { MockAuthRepository } from "@/config/__tests__/mocks/auth-repository.mock";
 
 describe("UpdateUserPassword", () => {
