+++ src/domain/user/use-cases/update-user.ts
@@ -1,16 +1,16 @@
 import { UpdateUserDto } from "@/domain/user/dtos/update-user.dto";
-import { AuthUserEntity } from "../entities/auth-user.entity";
-import { UserEntity } from "../entities/user.entity";
+import { AuthUserEntity } from "@/domain/entities/auth-user.entity";
+import { UserEntity } from "@/domain/entities/user.entity";
 import { AuthRepository } from "@/domain/auth/repositories/auth.repository";
 
-interface IUpdateUserPasswordUseCase {
+interface IUpdateUserUseCase {
   execute(dto: UpdateUserDto): Promise<UserEntity | AuthUserEntity>;
 }
 
-export class UpdateUserPassword implements IUpdateUserPasswordUseCase {
+export class UpdateUser implements IUpdateUserUseCase {
   constructor(private readonly repository: AuthRepository) {}
 
   async execute(dto: UpdateUserDto): Promise<UserEntity | AuthUserEntity> {
-    return this.repository.updateUserPassword(dto);
+    return this.repository.updateUser(dto);
   }
 }
